+++ azure_monitor/setup.cfg
@@ -27,8 +27,8 @@ package_dir=
     =src
 packages=find_namespace:
 install_requires =
-    opentelemetry-api ~= 0.7b1
-    opentelemetry-sdk ~= 0.7b1
+    opentelemetry-api == 0.9b0
+    opentelemetry-sdk == 0.9b0
     psutil >= 5.6.3
     requests ~= 2.0
