+++ azure_monitor/src/azure_monitor/sdk/auto_collection/live_metrics/exporter.py
@@ -4,7 +4,7 @@
 import logging
 import typing
 
-from opentelemetry.sdk.metrics import Counter, Observer
+from opentelemetry.sdk.metrics import ValueObserver, ValueRecorder
 from opentelemetry.sdk.metrics.export import (
     MetricRecord,
     MetricsExporter,
@@ -76,13 +76,19 @@ class LiveMetricsExporter(MetricsExporter):
         # Add metrics
         for metric_record in metric_records:
             value = 0
-            metric = metric_record.metric
-            if isinstance(metric, Counter):
-                value = metric_record.aggregator.checkpoint
-            elif isinstance(metric, Observer):
+            metric = metric_record.instrument
+            if isinstance(metric, ValueObserver):
+                # mmscl
                 value = metric_record.aggregator.checkpoint.last
-                if not value:
-                    value = 0
+            elif isinstance(metric, ValueRecorder):
+                # mmsc
+                value = metric_record.aggregator.checkpoint.count
+            else:
+                # sum or lv
+                value = metric_record.aggregator.checkpoint
+            if not value:
+                logger.warning("Value is none. Default to 0.")
+                value = 0
             envelope.metrics.append(
                 LiveMetric(name=metric.name, value=value, weight=1)
             )
